requirements.txt
asn1crypto==1.2.0
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto==2.49.0
boto3==1.10.9
botocore==1.13.9
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==1.11.8
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
docutils==0.15.2
future==0.18.2
idna==2.8
ijson==2.5.1
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
oscrypto==1.1.0
Pillow==6.2.0
pycparser==2.19
pycryptodomex==3.9.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
simplejson==3.17.0
six==1.12.0
snowflake-connector-python==2.0.3
uritemplate==3.0.0
urllib3==1.24.3
